Fix all CI failures: rustfmt, clippy, ruff, Python 3.10 compat, pytest-timeout
Rustfmt: formatted 20+ Rust files.

Clippy (30+ fixes):
- gen_range → random_range (rand 0.9 deprecation)
- len()==0 → is_empty(), iterator-based loops
- Dead code suppression on unused fields in candle/burn backends
- Literal simplification (8*1 → 8)

Ruff (43 fixes):
- 28 unused imports auto-fixed
- Added missing nn imports (dqn, sac, td3)
- Removed unused variables (calql, cql, dreamer, trpo)
- Renamed ambiguous 'l' → 'log' in builders
- Converted match→if/elif for Python 3.9 compat (runner.py)

Python 3.10 compat:
- typing.Self conditional import (added in 3.11) in 6 algorithm files

CI config:
- Added pytest-timeout to slow-tests job

+++ crates/rlox-candle/src/stochastic.rs
@@ -22,6 +22,7 @@ pub struct CandleStochasticPolicy {
     device: Device,
     act_dim: usize,
     lr: f64,
+    #[allow(dead_code)]
     rng: ChaCha8Rng,
 }
 
@@ -115,7 +116,7 @@ impl CandleStochasticPolicy {
 
         let y_sq = y_t.sqr().nn_err()?;
         let one = Tensor::ones_like(&y_sq).nn_err()?;
-        let correction = ((&one - &y_sq).nn_err()? + 1e-6 as f64)
+        let correction = ((&one - &y_sq).nn_err()? + 1e-6_f64)
             .nn_err()?
             .log()
             .nn_err()?
